+++ packages/menu/mwc-menu-base.ts
@@ -341,22 +341,11 @@ export abstract class MenuBase extends BaseElement {
     this.open = false;
   }
 
-  protected async _getUpdateComplete() {
-    return this.getUpdateComplete();
-  }
-
   // tslint:disable:ban-ts-ignore
   protected async getUpdateComplete() {
-    let result = false;
     await this._listUpdateComplete;
     // @ts-ignore
-    if (super.getUpdateComplete) {
-      // @ts-ignore
-      result = await super.getUpdateComplete();
-    } else {
-      // @ts-ignore
-      await super._getUpdateComplete();
-    }
+    const result = await super.getUpdateComplete();
     return result;
   }
   // tslint:enable:ban-ts-ignore