+++ packages/textfield/mwc-textfield-base.ts
@@ -627,21 +627,10 @@ export abstract class TextFieldBase extends FormElement {
     };
   }
 
-  protected async _getUpdateComplete() {
-    return this.getUpdateComplete();
-  }
-
   // tslint:disable:ban-ts-ignore
   protected async getUpdateComplete() {
-    let result = false;
     // @ts-ignore
-    if (super.getUpdateComplete) {
-      // @ts-ignore
-      result = await super.getUpdateComplete();
-    } else {
-      // @ts-ignore
-      await super._getUpdateComplete();
-    }
+    const result = await super.getUpdateComplete();
     await this._outlineUpdateComplete;
     return result;
   }